changed: put tvshow videodatabase methods in separate file
